+++ src/app/entry/entry_edit/entry_edit_controller.js
@@ -7,7 +7,7 @@ require('./editor.css');
 require('./editor');
 
 function EntryEditController($scope, $rootScope, $state, $stateParams, $timeout, $q,
-  Entry, Tag, Codekit, API, AuthService, DolphinService, toaster, Slug, $translate, $interval) {
+  Entry, Tag, Codekit, API, AuthService, DolphinService, toaster, Slug, $translate, $interval, ModalsService) {
   var payload;
   var tagsToCreate = [];
   var noneTagsCount = 0;
@@ -64,6 +64,7 @@ function EntryEditController($scope, $rootScope, $state, $stateParams, $timeout,
 
     if ($stateParams.entryId) {
       $scope.editing = true;
+      $scope.postInitial = false;
       // Load from cache if available
       if ($rootScope.cache.entry && $rootScope.cache.entry.get.id === $stateParams.entryId) {
         $scope.form = $rootScope.cache.entry;
@@ -80,6 +81,7 @@ function EntryEditController($scope, $rootScope, $state, $stateParams, $timeout,
           entry_id: $scope.form.get.id
         },
         function (data) {
+          $scope.postInitial = true;
           // Store original data
           oldData.originalPost = angular.copy(data);
           // If unsaved changes
@@ -117,6 +119,7 @@ function EntryEditController($scope, $rootScope, $state, $stateParams, $timeout,
           initAutoSave();
         },
         function () {
+          $scope.postInitial = false;
           $state.go("dash.entry-edit", { entryId: null });
           $translate(["OOPS", "ENTRY_GET_ERROR"]).then(function (translations) {
             toaster.error(translations.OOPS, translations.ENTRY_GET_ERROR);
@@ -490,6 +493,13 @@ function EntryEditController($scope, $rootScope, $state, $stateParams, $timeout,
       tagsToCreate = [];
     }
   });
+
+  /**
+   * @desc Open preview modal
+   */
+  $scope.preview = () => {
+    ModalsService.open("postPreview", "PreviewModalController", { URL: $scope.form.get.absolute_uri });
+  };
 
   /**
    * @desc Go to entries on entry removal